+++ admin/Views/Quanlydichvu.php
@@ -15,29 +15,7 @@ $this->layoutPath = "Layout.php";
 
             <button id="uploadfile" class="btn btn-success" type="button" title="Nhập"><a id="addnhanvien" href="index.php?controller=dichvu&action=create"><i class="fas fa-plus"></i>>
                 Tạo mới dịch vụ</a></button>
-            <!--
-                                <button id="uploadfile" class="btn btn-secondary btn-sm nhap-tu-file" type="button"
-                                    title="Nhập" onclick="myFunction(this)"><i class="fas fa-file-upload"></i>>
-                                    Tải từ file</button>
-
-
-
-                                <a class="btn btn-primary btn-sm print-file" type="button" title="In"
-                                    onclick="myApp.printTable()"><i class="fas fa-print"></i> In dữ liệu</a>
-
-
-                                <a class="btn btn-warning btn-sm print-file js-textareacopybtn" type="button"
-                                    title="Sao chép"><i class="fas fa-copy"></i> Sao chép</a>
-
-
-
-                                <a class="btn btn-success btn-sm" href="" title="In"><i class="fas fa-file-excel"></i>
-                                    Xuất Excel</a>
-
-
-                                <a class="btn btn-danger btn-sm pdf-file" type="button" title="In"
-                                    onclick="myFunction(this)"><i class="fas fa-file-pdf"></i> Xuất PDF</a>
- -->
+           
           </div>
           <div class="search mt-4 mb-4 input-group" style="width:50%">
             <button class="input-group-text btn btn-success"><i class="fa-solid fa-magnifying-glass"></i></button>
@@ -74,7 +52,7 @@ $this->layoutPath = "Layout.php";
                   <td><?php echo $row->ten_dichvu ?></td>
                   <td><?php echo $row->dateCreate ?></td>
                   <td>
-                    <a href="index.php?controller=dichvu&action=delete&id=<?php echo $row->id_dichvu ?>"> <button class="btn btn-danger">X</button></a>
+                    <a onclick="return confirm('Bạn có muốn xóa không')" href="index.php?controller=dichvu&action=delete&id=<?php echo $row->id_dichvu ?>"> <button class="btn btn-danger">X</button></a>
                     <a href="index.php?controller=dichvu&action=change&id=<?php echo $row->id_dichvu ?>"> <button class="btn btn-success"><i class="fas fa-edit"></i></button></a>
                     <a href="index.php?controller=dichvu&action=detail&id=<?php echo $row->id_dichvu ?>"><button class="btn btn-primary">Xem</button></a>
                   </td>
